[FIX] download SAMBA data
changed the  reader, so it correctly selectes the samba data from Meinen et al. Had to exclude the data from Kersale et al. because the ftp link is not valid and the file is not listed in the noaa database anymore . Will open another issue with this fix specifically. Closes Issue #92 .