[FIX] SAMBA data download
I tried a different way of accessing the ftp server and there the data showed up at the original place. Sorry for the confusion. The code still needs adjustment to correctly download the data, but with this commit I will push the Upper-Abyssal transport data in the repository so it can we worked with already and will fix the downloading issue in the next step.